Started script for reading in non-model related data.
